+++ src/commands/check.test.ts
@@ -6,6 +6,11 @@ import { mocked } from 'ts-jest/utils';
 import { ProfileId } from '../common/profile';
 import { check, CheckResult, formatHuman, formatJson } from '../logic/check';
 import { detectSuperJson } from '../logic/install';
+import {
+  MapFromMetadata,
+  ProfileFromMetadata,
+  ProviderFromMetadata,
+} from '../logic/publish.utils';
 import { MockStd, mockStd } from '../test/mock-std';
 import Check from './check';
 
@@ -25,22 +30,91 @@ describe('Check CLI command', () => {
   const profileId = 'starwars/character-information';
   const provider = 'swapi';
 
+  const version = '1.0.3';
+  // const mockProfileSource = 'mock profile source';
+  const mockMapSource = 'mock map source';
+
+  // const mockLocalProfileFrom: ProfileFromMetadata = {
+  //   kind: 'local',
+  //   source: mockProfileSource,
+  //   path: 'mock profile path'
+  // }
+
+  const mockLocalMapFrom: MapFromMetadata = {
+    kind: 'local',
+    source: mockMapSource,
+    path: 'mock map path',
+  };
+
+  const mockLocalProviderFrom: ProviderFromMetadata = {
+    kind: 'local',
+    path: 'mock provider path',
+  };
+
+  const mockRemoteProfileFrom: ProfileFromMetadata = {
+    kind: 'remote',
+    version,
+  };
+
+  const mockRemoteMapFrom: MapFromMetadata = {
+    kind: 'remote',
+    version,
+  };
+
+  // const mockRemoteProviderFrom: ProviderFromMetadata = {
+  //   kind: 'remote',
+  // }
+
   const mockResult: CheckResult[] = [
     {
-      kind: 'error',
-      message: 'first-error',
+      kind: 'profileMap',
+      provider,
+      profileFrom: mockRemoteProfileFrom,
+      mapFrom: mockRemoteMapFrom,
+      profileId,
+      issues: [
+        {
+          kind: 'error',
+          message: 'first-error',
+        },
+        {
+          kind: 'warn',
+          message: 'first-warn',
+        },
+        {
+          kind: 'error',
+          message: 'second-error',
+        },
+        {
+          kind: 'warn',
+          message: 'second-warn',
+        },
+      ],
     },
     {
-      kind: 'warn',
-      message: 'first-warn',
-    },
-    {
-      kind: 'error',
-      message: 'second-error',
-    },
-    {
-      kind: 'warn',
-      message: 'second-warn',
+      kind: 'mapProvider',
+      provider,
+      providerFrom: mockLocalProviderFrom,
+      mapFrom: mockLocalMapFrom,
+      profileId,
+      issues: [
+        {
+          kind: 'error',
+          message: 'first-error',
+        },
+        {
+          kind: 'warn',
+          message: 'first-warn',
+        },
+        {
+          kind: 'error',
+          message: 'second-error',
+        },
+        {
+          kind: 'warn',
+          message: 'second-warn',
+        },
+      ],
     },
   ];
   let stderr: MockStd;
@@ -67,7 +141,9 @@ describe('Check CLI command', () => {
       mocked(detectSuperJson).mockResolvedValue(undefined);
       await expect(
         Check.run(['--profileId', profileId, '--providerName', provider])
-      ).rejects.toEqual(new CLIError('Unable to check, super.json not found'));
+      ).rejects.toEqual(
+        new CLIError('❌ Unable to check, super.json not found')
+      );
     });
 
     it('throws when super.json not loaded correctly', async () => {
@@ -77,7 +153,9 @@ describe('Check CLI command', () => {
         .mockResolvedValue(err(new SDKExecutionError('test error', [], [])));
       await expect(
         Check.run(['--profileId', profileId, '--providerName', provider])
-      ).rejects.toEqual(new CLIError('Unable to load super.json: test error'));
+      ).rejects.toEqual(
+        new CLIError('❌ Unable to load super.json: test error')
+      );
     });
 
     it('throws error on invalid scan flag', async () => {
@@ -134,11 +212,11 @@ describe('Check CLI command', () => {
         ])
       ).rejects.toEqual(
         new CLIError(
-          'Invalid profile id: "U!0_" is not a valid lowercase identifier'
+          '❌ Invalid profile id: "U!0_" is not a valid lowercase identifier'
         )
       );
-      expect(detectSuperJson).toHaveBeenCalled();
-      expect(loadSpy).toHaveBeenCalled();
+      expect(detectSuperJson).not.toHaveBeenCalled();
+      expect(loadSpy).not.toHaveBeenCalled();
     }, 10000);
 
     it('throws error on invalid provider name', async () => {
@@ -156,9 +234,26 @@ describe('Check CLI command', () => {
           '-s',
           '3',
         ])
-      ).rejects.toEqual(new CLIError('Invalid provider name: "U!0_"'));
-      expect(detectSuperJson).toHaveBeenCalled();
-      expect(loadSpy).toHaveBeenCalled();
+      ).rejects.toEqual(new CLIError('❌ Invalid provider name: "U!0_"'));
+      expect(detectSuperJson).not.toHaveBeenCalled();
+      expect(loadSpy).not.toHaveBeenCalled();
+    }, 10000);
+
+    it('throws error on missing profile id when providerName is provided', async () => {
+      mocked(detectSuperJson).mockResolvedValue('.');
+      const loadSpy = jest
+        .spyOn(SuperJson, 'load')
+        .mockResolvedValue(ok(new SuperJson()));
+
+      await expect(
+        Check.run(['--providerName', provider, '-s', '3'])
+      ).rejects.toEqual(
+        new CLIError(
+          '❌ --profileId must be specified when using --providerName'
+        )
+      );
+      expect(detectSuperJson).not.toHaveBeenCalled();
+      expect(loadSpy).not.toHaveBeenCalled();
     }, 10000);
 
     it('throws error when profile Id not found in super.json', async () => {
@@ -177,7 +272,9 @@ describe('Check CLI command', () => {
           '3',
         ])
       ).rejects.toEqual(
-        new CLIError(`Profile id: "${profileId}" not found in super.json`)
+        new CLIError(
+          `❌ Unable to check, profile: "${profileId}" not found in super.json`
+        )
       );
       expect(detectSuperJson).toHaveBeenCalled();
       expect(loadSpy).toHaveBeenCalled();
@@ -207,7 +304,7 @@ describe('Check CLI command', () => {
         ])
       ).rejects.toEqual(
         new CLIError(
-          `Provider: "${provider}" not found in profile: "${profileId}" in super.json`
+          `❌ Unable to check, provider: "${provider}" not found in profile: "${profileId}" in super.json`
         )
       );
       expect(detectSuperJson).toHaveBeenCalled();
@@ -240,7 +337,9 @@ describe('Check CLI command', () => {
           '3',
         ])
       ).rejects.toEqual(
-        new CLIError(`Provider: "${provider}" not found in super.json`)
+        new CLIError(
+          `❌ Unable to check, provider: "${provider}" not found in super.json`
+        )
       );
       expect(detectSuperJson).toHaveBeenCalled();
       expect(loadSpy).toHaveBeenCalled();
@@ -276,17 +375,25 @@ describe('Check CLI command', () => {
           '-s',
           '3',
         ])
-      ).resolves.toBeUndefined();
+      ).rejects.toEqual(
+        new CLIError('❌ Command found 4 errors and 4 warnings')
+      );
       expect(detectSuperJson).toHaveBeenCalled();
       expect(loadSpy).toHaveBeenCalled();
       expect(check).toHaveBeenCalledWith(
         mockSuperJson,
-        {
-          id: ProfileId.fromScopeName('starwars', 'character-information'),
-          version: undefined,
-        },
-        provider,
-        { variant: undefined },
+        [
+          {
+            id: ProfileId.fromScopeName('starwars', 'character-information'),
+            maps: [
+              {
+                provider: provider,
+                variant: undefined,
+              },
+            ],
+            version: undefined,
+          },
+        ],
         { logCb: expect.anything(), warnCb: expect.anything() }
       );
       expect(stdout.output).toEqual('format-human\n');
@@ -324,17 +431,26 @@ describe('Check CLI command', () => {
           '3',
           '-q',
         ])
-      ).resolves.toBeUndefined();
+      ).rejects.toEqual(
+        new CLIError('❌ Command found 4 errors and 4 warnings')
+      );
+
       expect(detectSuperJson).toHaveBeenCalled();
       expect(loadSpy).toHaveBeenCalled();
       expect(check).toHaveBeenCalledWith(
         mockSuperJson,
-        {
-          id: ProfileId.fromScopeName('starwars', 'character-information'),
-          version: undefined,
-        },
-        provider,
-        { variant: undefined },
+        [
+          {
+            id: ProfileId.fromScopeName('starwars', 'character-information'),
+            maps: [
+              {
+                provider,
+                variant: undefined,
+              },
+            ],
+            version: undefined,
+          },
+        ],
         { logCb: undefined, warnCb: undefined }
       );
       expect(stdout.output).toEqual('format-human\n');
@@ -375,22 +491,236 @@ describe('Check CLI command', () => {
           '-q',
           '-j',
         ])
-      ).resolves.toBeUndefined();
+      ).rejects.toEqual(
+        new CLIError('❌ Command found 4 errors and 4 warnings')
+      );
       expect(detectSuperJson).toHaveBeenCalled();
       expect(loadSpy).toHaveBeenCalled();
       expect(check).toHaveBeenCalledWith(
         mockSuperJson,
-        {
-          id: ProfileId.fromScopeName('starwars', 'character-information'),
-          version: undefined,
-        },
-        provider,
-        { variant: undefined },
+        [
+          {
+            id: ProfileId.fromScopeName('starwars', 'character-information'),
+            maps: [
+              {
+                provider,
+                variant: undefined,
+              },
+            ],
+            version: undefined,
+          },
+        ],
+
         { logCb: undefined, warnCb: undefined }
       );
       expect(stdout.output).toContain('[{"kind": "error", "message": "test"}]');
       expect(formatJson).toHaveBeenCalledWith(mockResult);
       expect(formatHuman).not.toHaveBeenCalled();
+    });
+  });
+  describe('when preparing profiles to validation', () => {
+    const localProfile = 'local/profile';
+    const remoteProfile = 'remote/profile';
+    const localProvider = 'local-provider';
+    const remoteProvider = 'remote-provider';
+    const remoteProviderWithVarinat = 'remote-provider-with-variant';
+    const variant = 'variant';
+    const mockSuperJson = new SuperJson({
+      profiles: {
+        [localProfile]: {
+          file: 'profileFile',
+          providers: {
+            [localProvider]: {
+              file: 'mapPath',
+            },
+            [remoteProvider]: {},
+            [remoteProviderWithVarinat]: {
+              mapVariant: variant,
+            },
+          },
+        },
+        [remoteProfile]: {
+          version: '1.0.0',
+          providers: {
+            [localProvider]: {
+              file: 'mapPath',
+            },
+            [remoteProvider]: {},
+            [remoteProviderWithVarinat]: {
+              mapVariant: variant,
+            },
+          },
+        },
+      },
+      providers: {
+        [localProvider]: {},
+        [remoteProviderWithVarinat]: {},
+        [remoteProvider]: {},
+      },
+    });
+
+    it('prepares every local capability in super.json', async () => {
+      expect(Check.prepareProfilesToValidate(mockSuperJson)).toEqual([
+        {
+          id: ProfileId.fromId(localProfile),
+          maps: [
+            {
+              provider: localProvider,
+            },
+          ],
+        },
+      ]);
+    });
+
+    it('prepares specific profile id in super.json', async () => {
+      expect(
+        Check.prepareProfilesToValidate(mockSuperJson, localProfile)
+      ).toEqual([
+        {
+          id: ProfileId.fromId(localProfile),
+          maps: [
+            {
+              provider: localProvider,
+            },
+            {
+              provider: remoteProvider,
+            },
+            {
+              provider: remoteProviderWithVarinat,
+              variant,
+            },
+          ],
+        },
+      ]);
+
+      expect(
+        Check.prepareProfilesToValidate(mockSuperJson, remoteProfile)
+      ).toEqual([
+        {
+          id: ProfileId.fromId(remoteProfile),
+          maps: [
+            {
+              provider: localProvider,
+            },
+            {
+              provider: remoteProvider,
+            },
+            {
+              provider: remoteProviderWithVarinat,
+              variant,
+            },
+          ],
+          version: '1.0.0',
+        },
+      ]);
+    });
+
+    it('prepares specific profile and map in super.json', async () => {
+      expect(
+        Check.prepareProfilesToValidate(
+          mockSuperJson,
+          localProfile,
+          localProvider
+        )
+      ).toEqual([
+        {
+          id: ProfileId.fromId(localProfile),
+          maps: [
+            {
+              provider: localProvider,
+            },
+          ],
+        },
+      ]);
+
+      expect(
+        Check.prepareProfilesToValidate(
+          mockSuperJson,
+          remoteProfile,
+          localProvider
+        )
+      ).toEqual([
+        {
+          id: ProfileId.fromId(remoteProfile),
+          maps: [
+            {
+              provider: localProvider,
+            },
+          ],
+          version: '1.0.0',
+        },
+      ]);
+      expect(
+        Check.prepareProfilesToValidate(
+          mockSuperJson,
+          localProfile,
+          remoteProvider
+        )
+      ).toEqual([
+        {
+          id: ProfileId.fromId(localProfile),
+          maps: [
+            {
+              provider: remoteProvider,
+            },
+          ],
+        },
+      ]);
+
+      expect(
+        Check.prepareProfilesToValidate(
+          mockSuperJson,
+          remoteProfile,
+          remoteProvider
+        )
+      ).toEqual([
+        {
+          id: ProfileId.fromId(remoteProfile),
+          maps: [
+            {
+              provider: remoteProvider,
+            },
+          ],
+          version: '1.0.0',
+        },
+      ]);
+
+      expect(
+        Check.prepareProfilesToValidate(
+          mockSuperJson,
+          localProfile,
+          remoteProviderWithVarinat
+        )
+      ).toEqual([
+        {
+          id: ProfileId.fromId(localProfile),
+          maps: [
+            {
+              provider: remoteProviderWithVarinat,
+              variant,
+            },
+          ],
+        },
+      ]);
+
+      expect(
+        Check.prepareProfilesToValidate(
+          mockSuperJson,
+          remoteProfile,
+          remoteProviderWithVarinat
+        )
+      ).toEqual([
+        {
+          id: ProfileId.fromId(remoteProfile),
+          maps: [
+            {
+              provider: remoteProviderWithVarinat,
+              variant,
+            },
+          ],
+          version: '1.0.0',
+        },
+      ]);
     });
   });
 });