+++ src/logic/check.utils.ts
@@ -9,13 +9,16 @@ export async function findLocalProfileSource(
   superJson: SuperJson,
   profile: ProfileId,
   version?: string
-): Promise<string | undefined> {
+): Promise<{ path: string; source: string } | undefined> {
   //Check file property
   const profileSettings = superJson.normalized.profiles[profile.id];
   if (profileSettings !== undefined && 'file' in profileSettings) {
     const resolvedPath = superJson.resolvePath(profileSettings.file);
     if (await exists(resolvedPath)) {
-      return readFile(resolvedPath, { encoding: 'utf-8' });
+      return {
+        source: await readFile(resolvedPath, { encoding: 'utf-8' }),
+        path: resolvedPath,
+      };
     }
   }
 
@@ -28,7 +31,10 @@ export async function findLocalProfileSource(
     );
     const resolvedPath = superJson.resolvePath(path);
     if (await exists(resolvedPath)) {
-      return readFile(resolvedPath, { encoding: 'utf-8' });
+      return {
+        source: await readFile(resolvedPath, { encoding: 'utf-8' }),
+        path: resolvedPath,
+      };
     }
   } else {
     //Look for any version
@@ -47,7 +53,10 @@ export async function findLocalProfileSource(
       if (path) {
         const resolvedPath = superJson.resolvePath(joinPath(basePath, path));
         if (await exists(resolvedPath)) {
-          return readFile(resolvedPath, { encoding: 'utf-8' });
+          return {
+            source: await readFile(resolvedPath, { encoding: 'utf-8' }),
+            path: resolvedPath,
+          };
         }
       }
     }
@@ -60,7 +69,7 @@ export async function findLocalMapSource(
   superJson: SuperJson,
   profile: ProfileId,
   provider: string
-): Promise<string | undefined> {
+): Promise<{ path: string; source: string } | undefined> {
   //Check file property
   const profileSettings = superJson.normalized.profiles[profile.id];
   if (profileSettings !== undefined) {
@@ -68,7 +77,10 @@ export async function findLocalMapSource(
     if (profileProviderSettings && 'file' in profileProviderSettings) {
       const resolvedPath = superJson.resolvePath(profileProviderSettings.file);
       if (await exists(resolvedPath)) {
-        return readFile(resolvedPath, { encoding: 'utf-8' });
+        return {
+          source: await readFile(resolvedPath, { encoding: 'utf-8' }),
+          path: resolvedPath,
+        };
       }
     }
   }
@@ -79,7 +91,7 @@ export async function findLocalMapSource(
 export async function findLocalProviderSource(
   superJson: SuperJson,
   provider: string
-): Promise<ProviderJson | undefined> {
+): Promise<{ path: string; source: ProviderJson } | undefined> {
   //Check file property
   const providerSettings = superJson.normalized.providers[provider];
   if (
@@ -89,9 +101,12 @@ export async function findLocalProviderSource(
   ) {
     const resolvedPath = superJson.resolvePath(providerSettings.file);
     if (await exists(resolvedPath)) {
-      return JSON.parse(
-        await readFile(resolvedPath, { encoding: 'utf-8' })
-      ) as ProviderJson;
+      return {
+        path: resolvedPath,
+        source: JSON.parse(
+          await readFile(resolvedPath, { encoding: 'utf-8' })
+        ) as ProviderJson,
+      };
     }
   }