+++ src/logic/lint.test.ts
@@ -34,7 +34,7 @@ import {
   isValidHeader,
   isValidMapId,
   lint,
-  ProfileToLint,
+  ProfileToValidate,
 } from './lint';
 //Mock io
 jest.mock('../common/io', () => ({
@@ -64,6 +64,8 @@ jest.mock('@superfaceai/parser', () => ({
 }));
 
 describe('Lint logic', () => {
+  const mockMapPath = 'mockMapPath';
+  const mockProfilePath = 'mockProfilePath';
   afterEach(() => {
     jest.resetAllMocks();
   });
@@ -87,6 +89,7 @@ describe('Lint logic', () => {
   describe('when validating header', () => {
     let mockValidProfileHeader: ProfileHeaderStructure;
     let mocValidMapHeader: MapHeaderNode;
+
     beforeEach(() => {
       mockValidProfileHeader = {
         name: 'mockProfileHeader',
@@ -442,25 +445,30 @@ describe('Lint logic', () => {
         'character-information'
       );
       const mockSuperJson = new SuperJson();
-      const mockProfiles: ProfileToLint[] = [
+      const mockProfiles: ProfileToValidate[] = [
         {
           id: profile,
           maps: [
             {
               provider: 'swapi',
-              path: 'swapi path',
             },
             {
               provider: 'starwars',
-              path: 'starwars path',
             },
           ],
-          path: 'profile path',
         },
       ];
 
-      mocked(findLocalMapSource).mockResolvedValue(mockMapContent);
-      mocked(findLocalProfileSource).mockResolvedValue(mockProfileContent);
+      mocked(findLocalMapSource)
+        .mockResolvedValueOnce({ source: mockMapContent, path: 'swapi path' })
+        .mockResolvedValueOnce({
+          source: mockMapContent,
+          path: 'starwars path',
+        });
+      mocked(findLocalProfileSource).mockResolvedValue({
+        source: mockProfileContent,
+        path: mockProfilePath,
+      });
       mocked(parseProfile).mockReturnValue(mockProfileDocument);
       mocked(parseMap)
         .mockReturnValueOnce(mockMapDocument)
@@ -483,7 +491,7 @@ describe('Lint logic', () => {
         1,
         mockReportFn({
           kind: 'file',
-          path: 'profile path',
+          path: mockProfilePath,
           errors: [],
           warnings: [],
         })
@@ -503,7 +511,7 @@ describe('Lint logic', () => {
         3,
         mockReportFn({
           kind: 'compatibility',
-          profile: 'profile path',
+          profile: mockProfilePath,
           path: 'swapi path',
           errors: [
             {
@@ -533,7 +541,7 @@ describe('Lint logic', () => {
         5,
         mockReportFn({
           kind: 'compatibility',
-          profile: 'profile path',
+          profile: mockProfilePath,
           path: 'starwars path',
           errors: [],
           warnings: [],
@@ -547,7 +555,7 @@ describe('Lint logic', () => {
         'character-information'
       );
       const mockSuperJson = new SuperJson();
-      const mockProfiles: ProfileToLint[] = [
+      const mockProfiles: ProfileToValidate[] = [
         {
           id: profile,
           maps: [
@@ -560,12 +568,14 @@ describe('Lint logic', () => {
               variant: 'test',
             },
           ],
-          path: 'profile path',
         },
       ];
 
       mocked(findLocalMapSource).mockResolvedValue(undefined);
-      mocked(findLocalProfileSource).mockResolvedValue(mockProfileContent);
+      mocked(findLocalProfileSource).mockResolvedValue({
+        source: mockProfileContent,
+        path: mockProfilePath,
+      });
       mocked(parseProfile).mockReturnValue(mockProfileDocument);
       mocked(fetchMapAST)
         .mockResolvedValueOnce(mockMapDocument)
@@ -588,7 +598,7 @@ describe('Lint logic', () => {
         1,
         mockReportFn({
           kind: 'file',
-          path: 'profile path',
+          path: mockProfilePath,
           errors: [],
           warnings: [],
         })
@@ -598,7 +608,7 @@ describe('Lint logic', () => {
         2,
         mockReportFn({
           kind: 'compatibility',
-          profile: 'profile path',
+          profile: mockProfilePath,
           path: `${profile.id}.swapi.test@1.0.0`,
           errors: [
             {
@@ -617,7 +627,7 @@ describe('Lint logic', () => {
         3,
         mockReportFn({
           kind: 'compatibility',
-          profile: 'profile path',
+          profile: mockProfilePath,
           path: `${profile.id}.starwars.test@1.0.0`,
           errors: [],
           warnings: [],
@@ -631,7 +641,7 @@ describe('Lint logic', () => {
         'character-information'
       );
       const mockSuperJson = new SuperJson();
-      const mockProfiles: ProfileToLint[] = [
+      const mockProfiles: ProfileToValidate[] = [
         {
           id: profile,
           maps: [
@@ -722,7 +732,7 @@ describe('Lint logic', () => {
         'character-information'
       );
       const mockSuperJson = new SuperJson();
-      const mockProfiles: ProfileToLint[] = [
+      const mockProfiles: ProfileToValidate[] = [
         {
           id: profile,
           maps: [
@@ -739,7 +749,10 @@ describe('Lint logic', () => {
         },
       ];
 
-      mocked(findLocalProfileSource).mockResolvedValue(mockProfileContent);
+      mocked(findLocalProfileSource).mockResolvedValue({
+        source: mockProfileContent,
+        path: mockProfilePath,
+      });
       mocked(parseProfile).mockImplementation(() => {
         throw mockSyntaxErr;
       });
@@ -757,7 +770,7 @@ describe('Lint logic', () => {
       expect(writeElementSpy).toHaveBeenCalledWith(
         mockReportFn({
           kind: 'file',
-          path: 'starwars/character-information@1.0.0',
+          path: mockProfilePath,
           errors: [mockSyntaxErr],
           warnings: [],
         })
@@ -785,7 +798,7 @@ describe('Lint logic', () => {
         'character-information'
       );
       const mockSuperJson = new SuperJson();
-      const mockProfiles: ProfileToLint[] = [
+      const mockProfiles: ProfileToValidate[] = [
         {
           id: profile,
           maps: [
@@ -798,9 +811,15 @@ describe('Lint logic', () => {
         },
       ];
 
-      mocked(findLocalProfileSource).mockResolvedValue(mockProfileContent);
+      mocked(findLocalProfileSource).mockResolvedValue({
+        source: mockProfileContent,
+        path: mockProfilePath,
+      });
       mocked(parseProfile).mockReturnValue(mockProfileDocument);
-      mocked(findLocalMapSource).mockResolvedValue(mockMapContent);
+      mocked(findLocalMapSource).mockResolvedValue({
+        source: mockMapContent,
+        path: mockMapPath,
+      });
       mocked(parseMap).mockImplementation(() => {
         throw mockSyntaxErr;
       });
@@ -818,7 +837,7 @@ describe('Lint logic', () => {
         1,
         mockReportFn({
           kind: 'file',
-          path: 'starwars/character-information@1.0.0',
+          path: mockProfilePath,
           errors: [],
           warnings: [],
         })
@@ -827,7 +846,7 @@ describe('Lint logic', () => {
         2,
         mockReportFn({
           kind: 'file',
-          path: '',
+          path: mockMapPath,
           errors: [mockSyntaxErr],
           warnings: [],
         })